Improved page-zoom compatibility.
* A bit of body-padding, so the body text doesn't slam up against the
edge of the page
* Move the header left to compensate
* Zooming in on the hidden header text made it overflow out from behind
the image. The more I think about it the more I think that technique is
mostly to allow using a logo font without having to have it available on
the client machine. Well guess what: that header used verdana. So now
there's a small moustache-face image and a text-based text header.
* img { border: none} is kind've a shot in the dark. I know that some
browsers that I don't have draw borders around that header image. That
was fine when it was a banner in dimension but now it would step on the
text. This *should* do the trick? It might be time to get around to a
dang browser-testing VM.